add subscription button to alerts widgets
