+++ components/areas/AlertWidget.js
@@ -6,8 +6,15 @@ import debounce from 'lodash/debounce';
 // Redux
 import { connect } from 'react-redux';
 
+// Selectors
+import areaAlerts from 'selectors/user/areaAlerts';
+
 // Components
 import Map from 'components/ui/map/Map';
+
+import AreaSubscriptionModal from 'components/modal/AreaSubscriptionModal';
+import { toggleModal, setModalOptions } from 'redactions/modal';
+import { toggleTooltip } from 'redactions/tooltip';
 
 import MapControls from 'components/ui/map/MapControls';
 import ShareControl from 'components/ui/map/controls/ShareControl';
@@ -17,6 +24,8 @@ import BasemapControl from 'components/ui/map/controls/BasemapControl';
 import LayerManager from 'utils/layers/LayerManager';
 
 import { BASEMAPS, LABELS } from 'components/ui/map/constants';
+
+import DatasetService from 'services/DatasetService';
 
 // WRI components
 import {
@@ -31,10 +40,12 @@ class AlertWidget extends React.Component {
 
   constructor(props) {
     super(props);
-
-    const { dataset, user, subscription } = this.props;
+    const { dataset, subscription, user, id } = props;
+    const { areas } = user;
 
     this.state = {
+      area: areas.items.find(a => a.id === id),
+      subscription,
       zoom: 3,
       latLng: {
         lat: 0,
@@ -60,13 +71,40 @@ class AlertWidget extends React.Component {
     this.setState({ zoom, latLng });
   }, 250);
 
+  handleEditSubscription() {
+    const mode = this.state.subscription ? 'edit' : 'new';
+    const options = {
+      children: AreaSubscriptionModal,
+      childrenProps: {
+        area: this.state.area,
+        toggleModal: this.props.toggleModal,
+        onSubscriptionUpdated: this.handleSubscriptionUpdated,
+        onSubscriptionCreated: this.handleSubscriptionUpdated,
+        mode,
+        subscriptionDataset: true,
+        subscriptionType: true,
+        subscriptionThreshold: true
+      }
+    };
+    this.props.toggleModal(true);
+    this.props.setModalOptions(options);
+  }
+
   render() {
     const { dataset, user, subscription } = this.props;
     const layer = dataset.layer.find(l => l.default);
     const { zoom, latLng } = this.state;
     return (
       <div className="c-alerts-page__widget">
-        <h3 className="c-alerts-page__widget--heading">{layer && layer.name ? layer.name : 'not defined'}</h3>
+        <h2 className="c-alerts-page__widget--heading">{layer && layer.name ? layer.name : 'not defined'}</h2>
+
+        <button
+          className="c-btn -b"
+          onClick={() => this.handleEditSubscription()}
+        >
+          Edit Subscriptions
+        </button>
+
         {layer && <div className="c-alerts-page__graph">
           <Map
             mapConfig={{ zoom, latLng }}
@@ -110,7 +148,14 @@ AlertWidget.propTypes = {
 
 const mapStateToProps = state => ({
   user: state.user,
-  locale: state.common.locale
+  locale: state.common.locale,
+  alerts: areaAlerts(state)
 });
 
-export default connect(mapStateToProps, null)(AlertWidget);
+const mapDispatchToProps = {
+  toggleModal,
+  setModalOptions,
+  toggleTooltip
+};
+
+export default connect(mapStateToProps, mapDispatchToProps)(AlertWidget);
